refactor: code architecture improvements
