+++ package.json
@@ -12,6 +12,7 @@
   "author": "arpesh-gadekar",
   "license": "ISC",
   "dependencies": {
+    "@aws-sdk/client-s3": "^3.511.0",
     "axios": "^1.6.7",
     "bcrypt": "^5.1.1",
     "body-parser": "^1.20.2",
