--- src/config/constants.config.ts
@@ -1,1 +1,0 @@
-export const SALT_ROUNDS = 10;
