+++ src/middlewares/validateObjectId.ts
@@ -2,16 +2,15 @@ import { NextFunction, Request, Response } from "express";
 import { errorMessages } from "../constants/messages";
 import mongoose from "mongoose";
 
-export const checkForTeamId = (
+export const checkForObjectId = (
   req: Request,
   res: Response,
   next: NextFunction
 ) => {
-  const { teamId } = req.params;
-  if (!teamId)
-    return res.status(400).json({ message: errorMessages.TEAM_ID_REQUIRED });
+  const { id } = req.params;
+  if (!id) return res.status(400).json({ message: errorMessages.INVALID_ID });
 
-  if (!mongoose.Types.ObjectId.isValid(teamId))
+  if (!mongoose.Types.ObjectId.isValid(id))
     return res.status(400).json({ message: errorMessages.INVALID_ID });
 
   next();
